execute code in finally block to import features only if we successfully get a json or geojson. & a couple of todos

+++ das/mapping/admin.py
@@ -21,8 +21,6 @@ from django.template.response import TemplateResponse
 from django.utils.html import escape
 from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
-from arcgis2geojson import arcgis2geojson
-
 
 import mapping.models as models
 from core.openlayers import OSMGeoExtendedAdmin
@@ -471,6 +469,7 @@ class ArcgisConfigurationAdmin(admin.ModelAdmin):
             if member.type == "Feature Service" and member.title in items_for_demo:
                 title = member.title.replace(' ', '-')
                 logger.info(f'processing {title}')
+                file_ext = None
                 try:
                     data = member.layers[0].query().to_geojson
                     file_ext = 'geojson'
@@ -481,13 +480,17 @@ class ArcgisConfigurationAdmin(admin.ModelAdmin):
                     logger.info(f'Error reading from {member.title}', error)
                     messages.add_message(request, messages.ERROR, f'Could not read data from {member.title}')
                 finally:
-                    with open(f'./{title}.{file_ext}', 'w') as data_file:
-                        data_file.write(data)
-                        management.call_command(
-                            'importlayer', 'importspatialfile', data_file.name, id_field='GlobalID', source='ArcGIS'
-                        )
+                    if file_ext:
+                        with open(f'./{title}.{file_ext}', 'w') as data_file:
+                            data_file.write(data)
+                            # todo: Admin UI should have optional fields so user can specify the id_field name,
+                            #  source name and also name_field. These get passed as options to mgmt command
+                            management.call_command(
+                                'importlayer', 'importspatialfile', data_file.name, id_field='GlobalID', source='ArcGIS'
+                            )
                         os.remove(data_file.name)
-                    messages.add_message(request, messages.INFO, f'Features from {group.title} successfully loaded into ER')
+                        # todo: this adds a success message for each group member processed.
+                        messages.add_message(request, messages.INFO, f'Features from {group.title} successfully loaded into ER')
         logger.info('Returning from download_features')
 
         # Todo: Set a celery task that will update/download features periodically
